+++ test/setup_admin.dart
@@ -2,10 +2,10 @@
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
 import 'package:firebase_admin_interop/firebase_admin_interop.dart';
-import 'package:node_interop/node_interop.dart';
-import 'package:node_interop/test.dart';
+import 'package:node_interop/node.dart';
+import 'package:node_interop/util.dart';
 
-final Map<String, String> env = node.platform.environment;
+final Map<String, String> env = dartify(process.env);
 
 App initFirebaseApp() {
   if (!env.containsKey('FIREBASE_SERVICE_ACCOUNT_FILEPATH') ||
@@ -13,7 +13,6 @@ App initFirebaseApp() {
       !env.containsKey('FIREBASE_HTTP_BASE_URL')) {
     throw new StateError("Environment variables not set.");
   }
-  installNodeModules({"firebase-admin": "~4.2.1"});
 
   var admin = FirebaseAdmin.instance;
   return admin.initializeApp(